changed contact id to use big integer

+++ database/migrations/2019_10_11_010305_create_custom_attributes_table.php
@@ -15,7 +15,7 @@ class CreateCustomAttributesTable extends Migration
     {
         Schema::create('custom_attributes', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('contact_id');
+            $table->unsignedBigInteger('contact_id');
             $table->string('key');
             $table->text('value');
             $table->timestamps();
